+++ src/integrationTest/kotlin/org/jetbrains/intellij/platform/gradle/MultiModuleIntegrationTest.kt
@@ -9,6 +9,7 @@ import kotlin.io.path.notExists
 import kotlin.test.BeforeTest
 import kotlin.test.Ignore
 import kotlin.test.Test
+import kotlin.test.assertTrue
 
 private const val CUSTOM_RUN_IDE_TASK_NAME = "customRunIde"
 
@@ -31,6 +32,22 @@ class MultiModuleIntegrationTest : IntelliJPlatformIntegrationTestBase(
     fun `ext plugin uses base plugin from zip distribution`() {
         build(":ext:" + Tasks.BUILD_PLUGIN, projectProperties = defaultProjectProperties) {
 
+        }
+    }
+
+    @Test
+    fun `module project dependency is packaged to lib modules automatically`() {
+        val pluginDirectory = sandboxDirectory
+            .resolve("ext")
+            .resolve("$intellijPlatformType-$intellijPlatformVersion")
+            .resolve(Sandbox.PLUGINS)
+            .resolve("ext")
+
+        build(":ext:" + Tasks.PREPARE_SANDBOX, projectProperties = defaultProjectProperties) {
+            assertTrue(pluginDirectory.resolve("${Sandbox.Plugin.LIB_MODULES}/test.submodule.jar").exists())
+            assertTrue(pluginDirectory.resolve("${Sandbox.Plugin.LIB}/test.submodule.jar").notExists())
+            assertTrue(pluginDirectory.resolve("${Sandbox.Plugin.LIB}/dummy-0.1.2.jar").exists())
+            assertTrue(pluginDirectory.resolve("${Sandbox.Plugin.LIB}/raw-1.0.0.jar").exists())
         }
     }